backend/requirements.txt
flask
flask-cors
Flask-SQLAlchemy
Flask-Migrate
psycopg2
python-dotenv
